refactor(sql util): refactor out common sql utils
